Packaging constructor now takes prices as double and converts it to BigDecimal addPackaging_invalidArguments_throwsIllegalArgumentException test bugfixes

+++ src/main/java/domain/PackagingController.java
@@ -4,8 +4,6 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import persistence.PackagingDao;
 
-import java.math.BigDecimal;
-
 public class PackagingController {
 
     private final PackagingDao packagingDao;
@@ -22,7 +20,7 @@ public class PackagingController {
         return this.packagingList;
     }
 
-    public void addPackaging(String name, double width, double height, double length, BigDecimal price, String packagingType, boolean active) {
+    public void addPackaging(String name, double width, double height, double length, double price, String packagingType, boolean active) {
         if (packagingDao.exists(name, user.getSupplier().getSupplierId()))
             throw new IllegalArgumentException("Packaging type with this name already exists!");
         if (packagingType.isBlank() || packagingType.isEmpty())
